feat(arch): separate List from AdventureList
